requirements.txt
python-dotenv~=1.0.0
aiogram~=2.25.1
asgiref~=3.6.0
Django~=4.2